input: Fix repeated mute keys on some Lenovo ThinkPad models
On the model generations T410 up to the current T440 the mute button
on the keyboard is reported via the normal AT keyboard (not via the
"ThinkPad Extra Buttons" input device which is used on older models).

This key then has the following property:

- the LED on the key is toggled even with repeated keys
- only the first key press is reported from acpid2

This leads to problems if one assumes that every toggling of the mute
button is reported via the synthesized ACPI event button/mute.

This patch introduces a command line option --tpmutefix which counts
the repeated generation of the KEY_MUTE and creates an additional ACPI
event on the key release if the number of repeats is an odd number.